Fixed expansion of timeline columns in formatting/timeline.lisp
fixes #2365

The line cache mechanism of the timeline column only dropped cells at
the tail of the cache in order to get the cache down to the current
column width. However, when the column width changes (e.g. caused by
terminal window size change), it is also necessary to extend the cache
to bring it up to the current column width.

* formatting/timeline.lisp (adjust-cache! timeline): in addition to
  dropping cell at the tail of the cache, also expand the cache to the
  current width by adding cells, if necessary