More correct use of annotating converter in src/commands/bridge/participant.lisp
The annotating converter is used as an optimization that allows
listeners and informers working within the bridge to
forgo (de)serializing the event payload. If a transform needs to access
the payload, however, this optimization must not be applied. Therefore,
disable the optimization if a transform is present.

In the long run, transforms will be able to specify whether they need
access to the payload and (de)serialization will only be performed if
necessary, even in the presence of a transform.

* src/commands/bridge/participant.lisp
  (shared-initialize :after connection t): only configure annotating
  converter if no transform has been requested
* test/commands/bridge/command.lisp
  (test commands-bridge-command-root::smoke): test transforming events