Do not lock introspection participant in src/formatting/introspection/json.lisp
Some clients use the formatting style with an introspection database
object instead of an `remote-introspection' participant. In that case,
locking the introspection object is not possible.

To solve this, locking is now done around the `format-event' call in the
`command-execute' for the `introspect' command.

This, in turn, means that the delay behavior of the style can no longer
happen as part of `format-event'. As a workaround, setf style-database
now implements the delay behavior.

* src/commands/introspect.lisp (header): updated copyright
  (command-execute introspect): put database object into style; lock the
  introspection participant around the `format-event' call
* src/formatting/introspection/styles.lisp
  (format-event :before t delay-mixin t): removed; as a workaround,
  the waiting is now performed as part of setf style-database
  (setf style-database :after rt delay-mixin): new method; wait for
  introspection data before `format-event' is called
* src/formatting/introspection/json.lisp
  (format-event eql :dummy style-json t): do not lock the introspection
  object
* test/formatting/introspection/json.lisp
  (test rsb.formatting.introspection.json-root::smoke): lock the
  introspection participant; pass introspection database to style
* CMakeLists.txt: added test for introspect command with json style